requirements.txt
openai
rich
typer
python-dotenv
pytest
pytest-mock